MCloud-deploy-backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-random-queryset==0.1.3
django-seed==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==9.8.2
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
urllib3==1.26.7
whitenoise==5.3.0
